refactor: remove agent v1 implementation and standardize agent interface
