+++ packages/backend/src/ee/services/ai/utils/llmAsJudgeForTools.ts
@@ -21,7 +21,7 @@ import { JSONDiff } from 'autoevals';
 import { compact, differenceWith } from 'lodash';
 import { z } from 'zod';
 import { DbAiAgentToolCall } from '../../../database/entities/ai';
-import { defaultAgentOptions } from '../agents/agent';
+import { defaultAgentOptions } from '../agents/agentV2';
 import { getOpenaiGptmodel } from '../models/openai-gpt';
 
 const TOOL_NAME_TO_DB_TOOL_NAME = {
